requirements.txt
sentence-transformers
langchain
flask
python-dotenv
pinecone[grpc] 
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-google-genai
-e .